base 작업 진행
