+++ controllers/rankingControllers.js
@@ -1,4 +1,4 @@
-const poolAsyncAwait = require('./config/mysqlConfig.js').getMysqlPool2
+const poolAsyncAwait = require('../config/mysqlConfig.js').getMysqlPool2
 const logger = require("../util/logger.js")("ranking.js")
 const { USER_TABLE } = require('./user')
 const RankSchema = require('../models/ranking')
