+++ models/ploggging.js
@@ -1,0 +1,8 @@
+const mongoClient = require('../config/mongoConfig.js')
+
+const PloggingSchema = { };
+const MongoClient = require('../config/mongoConfig.js');
+const { ObjectId } = require('mongodb');
+
+module.exports = PloggingSchema;
+
