+++ models/users.js
@@ -51,11 +51,24 @@ const deleteUser = async(userId, t = null) => await User.destroy({
     where: {user_id: userId}
 }, { transaction: t});
 
+const updateUserPloggingData = async(updatedPloggingData, userId, t) => await User.update({
+        score_week: updatedPloggingData.score_week,
+        distance_week: updatedPloggingData.distance_week,
+        trash_week: updatedPloggingData.trash_week,
+        score_month: updatedPloggingData.score_month,
+        distance_month: updatedPloggingData.distance_month,
+        trash_month: updatedPloggingData.trash_month
+    },
+    { where: {user_id: userId}},
+    { transaction: t});
+
+
 module.exports = {
     findOneUser,
     createUser,
     updateUserName,
     updateUserImg,
     changeUserPassword,
-    deleteUser
+    deleteUser,
+    updateUserPloggingData
 }
